Move settings_t to the separeted file
Class settings_t and the helper function proxy_settings are moved from
proxy.(hpp|cpp) to ns_settings.(hpp|cpp).
