+++ src/ns_settings.cpp
@@ -1,0 +1,26 @@
+/*
+	Mediastorage-proxy is a HTTP proxy for mediastorage based on elliptics
+	Copyright (C) 2013-2015 Yandex
+
+	This program is free software; you can redistribute it and/or
+	modify it under the terms of the GNU General Public License
+	as published by the Free Software Foundation; either version 2
+	of the License, or (at your option) any later version.
+
+	This program is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+	GNU General Public License for more details.
+
+	You should have received a copy of the GNU General Public License
+	along with this program; if not, write to the Free Software
+	Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+*/
+
+#include "ns_settings.hpp"
+
+const elliptics::settings_t &
+elliptics::proxy_settings(const mastermind::namespace_state_t &ns_state) {
+	return static_cast<const settings_t &>(ns_state.settings().user_settings());
+}
+
